# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for indexing and querying blob transactions on Ethereum across multiple networks",
        "title": "Blob Indexer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/a-thomas-22/blob-indexer-api",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/blob/latest": {
            "get": {
                "description": "Retrieve the latest confirmed blob transactions from the blockchain",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blobs"
                ],
                "summary": "Get latest confirmed blobs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Network name or chain ID (default: first enabled network)",
                        "name": "network",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of blobs to return (default: 10, max: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of blobs to skip for pagination (default: 0, max: 10000)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by sender address",
                        "name": "from",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/api.BlobResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/blob/mempool": {
            "get": {
                "description": "Retrieve pending blob transactions from the mempool",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blobs"
                ],
                "summary": "Get pending blobs from mempool",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Network name or chain ID (default: first enabled network)",
                        "name": "network",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of blobs to return (default: 10, max: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of blobs to skip for pagination (default: 0, max: 10000)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by sender address",
                        "name": "from",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/api.BlobResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/blob/mempool/pressure": {
            "get": {
                "description": "Retrieve bounded aggregate pressure metrics for pending blob transactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blobs"
                ],
                "summary": "Get blob mempool pressure",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Network name or chain ID (default: first enabled network)",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.MempoolPressureResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/blob/pricing": {
            "get": {
                "description": "Retrieve current and historical blob pricing with utilization metrics and fork parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blobs"
                ],
                "summary": "Get blob pricing data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Network name or chain ID (default: first enabled network)",
                        "name": "network",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of recent blocks to include (default: 20, max: 100)",
                        "name": "blocks",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.PricingResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/blob/{txHash}": {
            "get": {
                "description": "Retrieve a specific blob transaction by its hash",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "blobs"
                ],
                "summary": "Get blob by transaction hash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Network name or chain ID (default: first enabled network)",
                        "name": "network",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Transaction hash",
                        "name": "txHash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.BlobResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "Blob not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
...
        "/dev/database": {
            "get": {
                "description": "Retrieve statistics about the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dev"
                ],
                "summary": "Get database statistics",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.DatabaseStats"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/dev/indexers": {
            "get": {
                "description": "Retrieve detailed metrics for all indexers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dev"
                ],
                "summary": "Get indexer metrics",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/api.IndexerMetrics"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/dev/logs": {
            "get": {
                "description": "Retrieve recent application logs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dev"
                ],
                "summary": "Get recent logs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of logs to return (default: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by log level (info, warn, error, debug)",
                        "name": "level",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/api.LogEntry"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
...
        "/dev/queries": {
            "get": {
                "description": "Retrieve statistics about database queries",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dev"
                ],
                "summary": "Get database query statistics",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of queries to return (default: 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/api.QueryStat"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/stats": {
            "get": {
                "description": "Retrieve statistics about blob transactions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Get blob statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Network name or chain ID (default: first enabled network)",
                        "name": "network",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.StatsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.BlobResponse": {
            "type": "object",
            "properties": {
                "base_fee_per_blob_gas": {
                    "type": "string"
                },
                "base_fee_per_blob_gas_gwei": {
                    "type": "string"
                },
                "blob_gas_used": {
                    "type": "integer"
                },
                "blob_index": {
                    "type": "integer"
                },
                "blob_size_bytes": {
                    "type": "integer"
                },
                "block_number": {
                    "type": "integer"
                },
                "block_url": {
                    "type": "string"
                },
                "confirmed": {
                    "type": "boolean"
                },
                "fee_cap_headroom_percent": {
                    "type": "string"
                },
                "fee_cap_headroom_wei": {
                    "type": "string"
                },
                "from_address": {
                    "type": "string"
                },
                "from_address_url": {
                    "type": "string"
                },
                "max_cost_wei": {
                    "type": "string"
                },
                "max_fee_per_blob_gas": {
                    "type": "string"
                },
                "max_fee_per_blob_gas_gwei": {
                    "type": "string"
                },
                "network_id": {
                    "type": "integer"
                },
                "network_name": {
                    "type": "string"
                },
                "realized_cost_wei": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                },
                "tip_per_blob_gas": {
                    "type": "string"
                },
                "tip_per_blob_gas_gwei": {
                    "type": "string"
                },
                "total_cost_eth": {
                    "type": "string"
                },
                "transaction_url": {
                    "type": "string"
                },
                "tx_hash": {
                    "type": "string"
                },
                "user_attribution": {
                    "type": "string"
                }
            }
        },
...
        "api.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
...
}